Enables program PDF generation
Adds the ability to generate and download a PDF version of the program,
including general information, flagship projects, and program items.

This enhancement provides a convenient way to share and archive the program
details in a readily accessible format.

Includes html2canvas and jsPDF dependencies for client-side PDF generation.
Also add utility functions to convert editorjs content to plain text and html